chore: split detecSignals
